% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{get_api_key}
\alias{get_api_key}
\title{Get the API key}
\usage{
get_api_key()
}
\value{
The current API key
}
\description{
Get the current API key set in the \code{OPENAUSTRALIA_API_KEY} environment
variable. If a variable has been set in this session using \code{\link[=set_api_key]{set_api_key()}}
then it will be returned, otherwise a key set in \code{~/.Renviron} will be
returned. If no key has been set then the function will stop with an error.
}
\details{
Apply for an API key at
\href{http://www.openaustralia.org.au/api/key}{the OpenAustralia site}.
}
