% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-comments.R
\name{get_comments}
\alias{get_comments}
\alias{getComments}
\title{Fetch comments left on OpenAustralia}
\usage{
get_comments(tidy = TRUE, date = NULL, search = NULL, user_id = NULL,
  pid = NULL, page = NULL, num = NULL)

getComments(tidy = TRUE, date = NULL, search = NULL, user_id = NULL,
  pid = NULL, page = NULL, num = NULL)
}
\arguments{
\item{tidy}{Convert the output list to a tidyverse-ready format}

\item{date}{Fetch the comments for this date}

\item{search}{Fetch the comments that contain this term}

\item{user_id}{Fetch the comments by a particular user ID}

\item{pid}{Fetch the comments made on a particular person ID
(Representative/Senator)}

\item{page}{Page of results to return}

\item{num}{Number of results to return}
}
\description{
Fetch comments left on OpenAustralia
}
