% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-debates.R
\name{get_debates}
\alias{get_debates}
\alias{getDebates}
\title{Fetch Debates}
\usage{
get_debates(tidy = TRUE, type = NULL, date = NULL, search = NULL,
  person = NULL, gid = NULL, order = NULL, page = NULL, num = NULL)

getDebates(tidy = TRUE, type = NULL, date = NULL, search = NULL,
  person = NULL, gid = NULL, order = NULL, page = NULL, num = NULL)
}
\arguments{
\item{tidy}{Convert the output list to a tidyverse-ready format}

\item{type}{One of "representatives" or "senate"}

\item{date}{Fetch the debates for this date}

\item{search}{Fetch the debates that contain this term}

\item{person}{Fetch the debates by a particular person ID}

\item{gid}{Fetch the speech or debate that matches this GID}

\item{order}{d for date ordering, r for relevance ordering}

\item{page}{Page of results to return}

\item{num}{Number of results to return}
}
\description{
This includes oral questions.
}
