% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-divisions.R
\name{get_divisions}
\alias{get_divisions}
\alias{getDivisions}
\title{Fetch a list of electoral divisions}
\usage{
get_divisions(postcode = NULL, date = NULL, search = NULL)

getDivisions(postcode = NULL, date = NULL, search = NULL)
}
\arguments{
\item{postcode}{Fetch the list of electoral divisions that are within the
given postcode (there can be more than one)}

\item{date}{ISO-style date, e.g. "1990-01-02", to fetch the list of electoral
divisions as it was on this date}

\item{search}{Fetch the list of electoral divisions that match this search
string}
}
\description{
Fetch a list of electoral divisions
}
