% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-hansard.R
\name{get_hansard}
\alias{get_hansard}
\alias{getHansard}
\title{Fetch all Hansard}
\usage{
get_hansard(search = NULL, person = NULL, order = NULL, page = NULL,
  num = NULL)

getHansard(search = NULL, person = NULL, order = NULL, page = NULL,
  num = NULL)
}
\arguments{
\item{search}{Fetch the data that contain this term}

\item{person}{Fetch the data by a particular person ID}

\item{order}{d for date ordering, r for relevance ordering, p for use by
person}

\item{page}{Page of results to return}

\item{num}{Number of results to return}
}
\description{
Fetch all Hansard
}
