% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-representative.R
\name{get_representative}
\alias{get_representative}
\alias{getRepresentative}
\title{Fetch a particular member of the House of Representatives}
\usage{
get_representative(tidy = TRUE, id = NULL, division = NULL,
  always_return = NULL)

getRepresentative(tidy = TRUE, id = NULL, division = NULL,
  always_return = NULL)
}
\arguments{
\item{tidy}{Convert the output list to a tidyverse-ready format}

\item{id}{If you know the person ID for the member you want (returned from
get_representatives or elsewhere), this will return data for that person}

\item{division}{The name of an electoral division; OpenAustralia will try and
work it out from whatever you give them}

\item{always_return}{For the division option, sets whether to always try and
return a Representative, even if the seat is currently vacant}
}
\description{
Fetch a particular member of the House of Representatives
}
