% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-representatives.R
\name{get_representatives}
\alias{get_representatives}
\alias{getRepresentatives}
\title{Fetch a list of members of the House of Representatives}
\usage{
get_representatives(tidy = TRUE, postcode = NULL, date = NULL,
  party = NULL, search = NULL)

getRepresentatives(tidy = TRUE, postcode = NULL, date = NULL,
  party = NULL, search = NULL)
}
\arguments{
\item{tidy}{Convert the output list to a tidyverse-ready format}

\item{postcode}{Fetch the list of Representatives whose electoral division
lies within the postcode (there may be more than one)}

\item{date}{Fetch the list of members of the House of Representatives as it
was on this date}

\item{party}{Fetch the list of Representatives from the given party}

\item{search}{Fetch the list of Representatives that match this search string
in their name}
}
\description{
Fetch a list of members of the House of Representatives
}
