% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-senator.R
\name{get_senator}
\alias{get_senator}
\alias{getSenator}
\title{Fetch a particular Senator}
\usage{
get_senator(tidy = TRUE, id = NULL)

getSenator(tidy = TRUE, id = NULL)
}
\arguments{
\item{tidy}{Convert the output list to a tidyverse-ready format}

\item{id}{If you know the person ID for the Senator you want, this will
return data for that person}
}
\description{
Fetch a particular Senator
}
