% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-senators.R
\name{get_senators}
\alias{get_senators}
\alias{getSenators}
\title{Fetch a list of Senators}
\usage{
get_senators(date = NULL, party = NULL, state = NULL, search = NULL)

getSenators(date = NULL, party = NULL, state = NULL, search = NULL)
}
\arguments{
\item{date}{ISO-style date, e.g. "1990-01-02", to fetch the list of Senators
as it was on this date}

\item{party}{Fetch the list of Senators from the given party}

\item{state}{Fetch the list of Senators from the given state (NSW, Tasmania,
WA, Queensland, Victoria, SA, NT, ACT)}

\item{search}{Fetch the list of Senators that match this search string in
their name}
}
\description{
Fetch a list of Senators
}
