% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{set_api_key}
\alias{set_api_key}
\title{Set the API key}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{a new API key}
}
\value{
Nothing
}
\description{
This function manually sets an API key that will be used in the current
session. Call this function if you haven't set an
\code{OPENAUSTRALIA_API_KEY} environment variable in \code{~/.Renviron} or
if you want to use a different key throughout your current session.
}
\details{
Apply for an API key at
\href{http://www.openaustralia.org.au/api/key}{the OpenAustralia site}.
}
