% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-debates-by-date.R
\name{tibble_debates_by_date}
\alias{tibble_debates_by_date}
\title{Convert a list of debates returned from an API search by date into a nested
tibble}
\usage{
tibble_debates_by_date(data)
}
\arguments{
\item{data}{A nested list (list of lists) with each item in the top-level
list representing a debate.}
}
\value{
A tibble with each debate represented by one row.
}
\description{
Convert a list of debates returned from an API search by date into a nested
tibble
}
