% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-members.R
\name{tibble_members}
\alias{tibble_members}
\title{Convert a sub-list of members into a nested tibble}
\usage{
tibble_members(data, has_office)
}
\arguments{
\item{data}{A nested list (list of lists) with each item in the top-level
list representing a member. All members must have office data or must not
have office data}

\item{has_office}{A boolean value indicating whether the sub-list contains
office data}
}
\value{
A tibble with each member represented by one row
}
\description{
Convert a sub-list of members into a nested tibble
}
