% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-members.R
\name{tibble_office}
\alias{tibble_office}
\title{Convert a list of offices held by a member into a tibble, or create an
empty tibble if the member has no office data.}
\usage{
tibble_office(data)
}
\arguments{
\item{data}{A list with each item representing an office held by the member.}
}
\value{
A tibble with each office represented by one row.
}
\description{
Convert a list of offices held by a member into a tibble, or create an
empty tibble if the member has no office data.
}
