% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-debates-by-date.R
\name{tibble_subs}
\alias{tibble_subs}
\title{Convert a list of sub-debates into a tibble.}
\usage{
tibble_subs(data)
}
\arguments{
\item{data}{A list with each item representing a sub-debate.}
}
\value{
A tibble with each sub-debate represented by one row.
}
\description{
Convert a list of sub-debates into a tibble.
}
