% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-debates-by-date.R
\name{tidy_debates_by_date}
\alias{tidy_debates_by_date}
\title{Process nested list data of debates returned from an API search by date and
return a tidy nested tibble}
\usage{
tidy_debates_by_date(data)
}
\arguments{
\item{data}{A nested list (list of lists) with each item in the top-level
list representing a debate}
}
\value{
A tibble with each debate represented by one row.
}
\description{
Process nested list data of debates returned from an API search by date and
return a tidy nested tibble
}
