% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-divisions.R
\name{tidy_divisions}
\alias{tidy_divisions}
\title{Process list data of divisions and return a tidy tibble}
\usage{
tidy_divisions(data)
}
\arguments{
\item{data}{A nested list (list of lists) with each item in the top-level
list representing a division}
}
\value{
A tibble with each division represented by one row.
}
\description{
Process list data of divisions and return a tidy tibble
}
