% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-members.R
\name{tidy_members}
\alias{tidy_members}
\title{Process nested list data of members and return a tidy nested tibble}
\usage{
tidy_members(data)
}
\arguments{
\item{data}{A nested list (list of lists) with each item in the top-level
list representing a member}
}
\value{
A tibble with each member represented by one row.
}
\description{
Process nested list data of members and return a tidy nested tibble
}
