% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-members.R
\name{type_convert_members}
\alias{type_convert_members}
\title{Convert the column types for a tibble of members from strings (character) to
integers, dates, etc., and replace empty strings and dummy dates with NA
values.}
\usage{
type_convert_members(data)
}
\arguments{
\item{data}{A tibble with each member represented by one row.}
}
\value{
A tibble with column types converted and NA values applied.
}
\description{
Convert the column types for a tibble of members from strings (character) to
integers, dates, etc., and replace empty strings and dummy dates with NA
values.
}
