% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-members.R
\name{tibble_members}
\alias{tibble_members}
\title{Convert a list of members into a nested tibble}
\usage{
tibble_members(data)
}
\arguments{
\item{data}{A nested list (list of lists) with each item in the top-level
list representing a member.}
}
\value{
A tibble with each member represented by one row.
}
\description{
Convert a list of members into a nested tibble
}
